Refactor: SignupRepository and SignupRepositoryImpl packages
